binary: Format without dropping zeros
The hex string formatting of hash_image binary (which is the only
example of hex-formatting in the code base) used to elide leading zeros,
which makes the resulting hashes hard to compare against other
implementations' output values, and can be confusing when trying to
understand the resulting hash's size.

Casing is set to lower-case hex as that appears to be more common in
other hash implementations.