Fix port to rustdct 0.7
process_dct2_with_scratch() operates inplace now and the second
parameter is the scratch space, not the output.

Also the size of the scratch space depends on the algorithm that is
selected at runtime, and also only depends on the max(width, height)
instead of width*height so less scratch space can be allocated now.